bulk useradd example
This is the the script I used to add bulk users to my active directory via .csv and powershell. I tried using a few different scripts and modifying/debugging them but this one worked the best for me. Credit to JitenSH at (https://community.spiceworks.com/scripts/show/3682-bulk-create-active-directory-users-powershell-with-logs-less-rows-in-csv) for the template.